+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## 1.1.0 - ?? May 2020
+
+Breaking changes:
+- (None)
+
+New features:
+- Added TypeExtensions.GetPropertiesWithAttribute
+- Added TypeExtensions.GetEnumProperties
+- Added TypeExtensions.GetPropertiesOfType
+
+Bug fixes:
+- (None)
+
 ## 1.0.0 - 07 May 2020
 
 Initial version.